+++ Common/src/main/java/net/potionstudios/biomeswevegone/world/level/levelgen/structure/processor/processors/FruitBlockProcessor.java
@@ -58,6 +58,6 @@ public class FruitBlockProcessor extends StructureProcessor {
 
 	@Override
 	protected @NotNull StructureProcessorType<?> getType() {
-		return BWGCustomStructureProcessors.FRUIT_BLOCK_PROCESSOR;
+		return BWGCustomStructureProcessors.FRUIT_BLOCK_PROCESSOR.get();
 	}
 }