+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.2.1-Beta
 - Fix Oddion Foods having wrong Nutrition Values
 - Remove need for Explicit BlockPredicate Registration Method
+- Update Custom Structure Processor Registration
 
 # 1.2.0-Beta
 - Update/Add To French Language File (FR_FR)